Added dummy GUI to simulate grasp stability estimates
The python GUI offers a simple way to simulate a grasp stability estimator. It advertises a ROS service with which new gripper contexts are registered, and let's a user send arbitrary stability estimates over a dedicated topic. This GUI works with the CRAM macro `with-grasp-stability-awareness' and implements the protocol that was agreed earlier on that topic.